bot/requirements.txt
discord.py
SQLAlchemy
requests